% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\alias{blblm}
\alias{blblm-package}
\title{blblm: little bootstrap and regression}
\description{
It returns coefficients, models for regression, glm, both with or without parrelel.

It returns coefficients, models for regression, glm, both with or without parrelel.
}
\details{
Linear Regression with Little Bag of Bootstraps

Regression models with Little Bag of Bootstraps
}
