% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{linear regression using LBB}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data frame}

\item{m}{splitted data to m parts, default 10 splits}

\item{B}{numbers of bootstrap, default 5000}
}
\description{
linear regression using LBB
}
