% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{coefficients for bootstrap lm}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{fit}

\item{...}{arguments}
}
\description{
coefficients for bootstrap lm
}
