% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{glm for each bootstrap, specifying frequency}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{freqs}{weights for each linear regressor}

\item{family}{glm family}
}
\description{
glm for each bootstrap, specifying frequency
}
