% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{compute glm for each bootstrap}
\usage{
glm_each_boot(formula, data, n, family)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{n}{how many vectors to draw and use}

\item{family}{glm family}
}
\description{
compute glm for each bootstrap
}
