% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{lm for each subsample}
\usage{
glm_each_subsample(formula, data, n, B, family)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{n}{how many vectors to use}

\item{B}{numbers of bootstrap}

\item{family}{glm family}
}
\description{
lm for each subsample
}
