% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{lm for each bootstrap, specifying frequency}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{freqs}{weights for each linear regressor}
}
\description{
lm for each bootstrap, specifying frequency
}
