% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute lm for each bootstrap}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{n}{how many vectors to draw and use}
}
\description{
compute lm for each bootstrap
}
