% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{lm for each subsample}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{n}{how many vectors to use}

\item{B}{numbers of bootstrap}
}
\description{
lm for each subsample
}
