% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blbglm}
\alias{par_blbglm}
\title{generalized linear regression with parrel}
\usage{
par_blbglm(formula, data, m = 10, B = 5000, family)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{m}{number of splits}

\item{B}{number of bootstraps}

\item{family}{some glm family to use}
}
\description{
generalized linear regression with parrel
}
