% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blblm}
\alias{par_blblm}
\title{linear regression using LLB, with parallelization}
\usage{
par_blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression model}

\item{data}{data frame}

\item{m}{number of splits}

\item{B}{number of bootstraps}
}
\description{
linear regression using LLB, with parallelization
}
