% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{complute sigma for bootstrap regression fit}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{LBB regression}

\item{confidence}{logical/boolean value}

\item{level}{overall intented confidence level for sigma's CI}

\item{...}{other customized arguments}
}
\description{
complute sigma for bootstrap regression fit
}
